refactor: seperate detector and viewer
