+++ ninshiki_yolo/ninshiki_yolo_viewer.py
@@ -18,44 +18,23 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import argparse
 import rclpy
 from rclpy.node import Node
 
-from ninshiki_yolo.detector.node.detection import Detection
-from ninshiki_yolo.detector.node.detector_node import DetectorNode
 from ninshiki_yolo.viewer.node.viewer import Viewer
 from ninshiki_yolo.viewer.node.viewer_node import ViewerNode
-from ninshiki_yolo.node.ninshiki_yolo_node import NinshikiYoloNode
 
 def main(args=None):
     rclpy.init(args=args)
 
-    parser = argparse.ArgumentParser()
-    parser.add_argument('topic', help='specify topic name to subscribe')
-    # parser.add_argument('config', help='specify model configuration')
-    # parser.add_argument('names', help='specify class file name')
-    # parser.add_argument('weights', help='specify model weights')
-    parser.add_argument('--postprocess', help='show detection result', type=int,
-                        choices=[0, 1], default=0)
-    parser.add_argument('--GPU', help='if we chose the computation using GPU', type=int,
-                        choices=[0, 1], default=0)
-    parser.add_argument('--MYRIAD', help='if we chose the computation using Compute Stick', type=int,
-                        choices=[0, 1], default=0)
-    arg = parser.parse_args()
+    node = Node("ninshiki_yolo_viewer")
 
-    node = Node("ninshiki_yolo")
-    ninshiki_yolo_node = NinshikiYoloNode(node, arg.topic, arg.postprocess)
-
-    detection = Detection(gpu = arg.GPU, myriad = arg.MYRIAD)
     viewer = Viewer()
+    viewer_node = ViewerNode(node, "camera/image", "ninshiki_yolo/detection", viewer)
 
-    ninshiki_yolo_node.set_detector(detection)
-    ninshiki_yolo_node.set_viewer(viewer)
+    rclpy.spin(viewer_node.node)
 
-    rclpy.spin(ninshiki_yolo_node.node)
-
-    ninshiki_yolo_node.destroy_node()
+    viewer_node.destroy_node()
     rclpy.shutdown()
